+++ TShockAPI/Bouncer.cs
@@ -37,14 +37,18 @@ namespace TShockAPI
 	internal sealed class Bouncer
 	{
 		internal Handlers.SendTileSquareHandler STSHandler { get; set; }
-		internal Handlers.LandGolfBallInCupHandler LandGolfBallInCupHandler { get; set; }
+		internal Handlers.NetModules.NetModulePacketHandler NetModuleHandler { get; set; }
+    internal Handlers.LandGolfBallInCupHandler LandGolfBallInCupHandler { get; set; }
 
 		/// <summary>Constructor call initializes Bouncer and related functionality.</summary>
 		/// <returns>A new Bouncer.</returns>
 		internal Bouncer()
 		{
 			STSHandler = new Handlers.SendTileSquareHandler();
-			GetDataHandlers.SendTileSquare += STSHandler.OnReceiveSendTileSquare;
+			GetDataHandlers.SendTileSquare += STSHandler.OnReceive;
+
+			NetModuleHandler = new Handlers.NetModules.NetModulePacketHandler();
+			GetDataHandlers.ReadNetModule += NetModuleHandler.OnReceive;
 
 			LandGolfBallInCupHandler = new Handlers.LandGolfBallInCupHandler();
 			GetDataHandlers.LandGolfBallInCup += LandGolfBallInCupHandler.OnLandGolfBallInCup;