+++ TShockAPI/GetDataHandlers.cs
@@ -2002,6 +2002,40 @@ namespace TShockAPI
 			return args.Handled;
 		}
 
+		/// <summary>
+		/// Used when a net module is loaded
+		/// </summary>
+		public class ReadNetModuleEventArgs : GetDataHandledEventArgs
+		{
+			/// <summary>
+			/// The type of net module being loaded
+			/// </summary>
+			public NetModuleType ModuleType { get; set; }
+		}
+
+		/// <summary>
+		/// Called when a net module is received
+		/// </summary>
+		public static HandlerList<ReadNetModuleEventArgs> ReadNetModule = new HandlerList<ReadNetModuleEventArgs>();
+
+		private static bool OnReadNetModule(TSPlayer player, MemoryStream data, NetModuleType moduleType)
+		{
+			if (ReadNetModule == null)
+			{
+				return false;
+			}
+
+			var args = new ReadNetModuleEventArgs
+			{
+				Player = player,
+				Data = data,
+				ModuleType = moduleType
+			};
+
+			ReadNetModule.Invoke(null, args);
+			return args.Handled;
+		}
+
 		#endregion
 
 		private static bool HandlePlayerInfo(GetDataHandlerArgs args)
@@ -2259,11 +2293,10 @@ namespace TShockAPI
 
 			else if ((Main.ServerSideCharacter) && (args.Player.sX > 0) && (args.Player.sY > 0) && (args.TPlayer.SpawnX > 0) && ((args.TPlayer.SpawnX != args.Player.sX) && (args.TPlayer.SpawnY != args.Player.sY)))
 			{
-
 				args.Player.sX = args.TPlayer.SpawnX;
 				args.Player.sY = args.TPlayer.SpawnY;
 
-				if (((Main.tile[args.Player.sX, args.Player.sY - 1].active() && Main.tile[args.Player.sX, args.Player.sY - 1].type == 79)) && (WorldGen.StartRoomCheck(args.Player.sX, args.Player.sY - 1)))
+				if (((Main.tile[args.Player.sX, args.Player.sY - 1].active() && Main.tile[args.Player.sX, args.Player.sY - 1].type == TileID.Beds)) && (WorldGen.StartRoomCheck(args.Player.sX, args.Player.sY - 1)))
 				{
 					args.Player.Teleport(args.Player.sX * 16, (args.Player.sY * 16) - 48);
 					TShock.Log.ConsoleDebug("GetDataHandlers / HandleSpawn force teleport phase 1 {0}", args.Player.Name);
@@ -2272,7 +2305,7 @@ namespace TShockAPI
 
 			else if ((Main.ServerSideCharacter) && (args.Player.sX > 0) && (args.Player.sY > 0))
 			{
-				if (((Main.tile[args.Player.sX, args.Player.sY - 1].active() && Main.tile[args.Player.sX, args.Player.sY - 1].type == 79)) && (WorldGen.StartRoomCheck(args.Player.sX, args.Player.sY - 1)))
+				if (((Main.tile[args.Player.sX, args.Player.sY - 1].active() && Main.tile[args.Player.sX, args.Player.sY - 1].type == TileID.Beds)) && (WorldGen.StartRoomCheck(args.Player.sX, args.Player.sY - 1)))
 				{
 					args.Player.Teleport(args.Player.sX * 16, (args.Player.sY * 16) - 48);
 					TShock.Log.ConsoleDebug("GetDataHandlers / HandleSpawn force teleport phase 2 {0}", args.Player.Name);
@@ -2979,21 +3012,21 @@ namespace TShockAPI
 			if (bosses.Contains(thingType) && !args.Player.HasPermission(Permissions.summonboss))
 			{
 				TShock.Log.ConsoleDebug("GetDataHandlers / HandleSpawnBoss rejected boss {0} {1}", args.Player.Name, thingType);
-				args.Player.SendErrorMessage("You don't have permission to summon a boss.");
+				args.Player.SendErrorMessage("You do not have permission to summon a boss.");
 				return true;
 			}
 
 			if (invasions.Contains(thingType) && !args.Player.HasPermission(Permissions.startinvasion))
 			{
 				TShock.Log.ConsoleDebug("GetDataHandlers / HandleSpawnBoss rejected invasion {0} {1}", args.Player.Name, thingType);
-				args.Player.SendErrorMessage("You don't have permission to start an invasion.");
+				args.Player.SendErrorMessage("You do not have permission to start an invasion.");
 				return true;
 			}
 
 			if (pets.Contains(thingType) && !args.Player.HasPermission(Permissions.spawnpets))
 			{
 				TShock.Log.ConsoleDebug("GetDataHandlers / HandleSpawnBoss rejected pet {0} {1}", args.Player.Name, thingType);
-				args.Player.SendErrorMessage("You don't have permission to spawn pets.");
+				args.Player.SendErrorMessage("You do not have permission to spawn pets.");
 				return true;
 			}
 
@@ -3276,160 +3309,12 @@ namespace TShockAPI
 		private static bool HandleLoadNetModule(GetDataHandlerArgs args)
 		{
 			short moduleId = args.Data.ReadInt16();
-			if (moduleId == (int)NetModulesTypes.CreativePowers)
+
+			if (OnReadNetModule(args.Player, args.Data, (NetModuleType)moduleId))
 			{
-				CreativePowerTypes powerId = (CreativePowerTypes)args.Data.ReadInt16();
-				switch (powerId)
-				{
-					case CreativePowerTypes.FreezeTime:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_timefreeze))
-							{
-								args.Player.SendErrorMessage("You don't have permission to freeze the time of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.SetDawn:
-					case CreativePowerTypes.SetNoon:
-					case CreativePowerTypes.SetDusk:
-					case CreativePowerTypes.SetMidnight:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_timeset))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the time of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.Godmode:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_godmode))
-							{
-								args.Player.SendErrorMessage("You don't have permission to toggle godmode!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.WindStrength:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_windstrength))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the wind strength of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.RainStrength:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_rainstrength))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the rain strength of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.TimeSpeed:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_timespeed))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the time speed of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.RainFreeze:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_rainfreeze))
-							{
-								args.Player.SendErrorMessage("You don't have permission to freeze the rain strength of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.WindFreeze:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_windfreeze))
-							{
-								args.Player.SendErrorMessage("You don't have permission to freeze the wind strength of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.IncreasePlacementRange:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_placementrange))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the tile placement range of your character!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.WorldDifficulty:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_setdifficulty))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the world difficulty of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.BiomeSpreadFreeze:
-						{
-							if (!args.Player.HasPermission(Permissions.journey_biomespreadfreeze))
-							{
-								args.Player.SendErrorMessage("You don't have permission to freeze the biome spread of the server!");
-								return true;
-							}
-							break;
-						}
-					case CreativePowerTypes.SetSpawnRate:
-						{
-							// This is a monkeypatch because the 1.4.0.4 seemingly at random sends NPC spawn rate changes even outside of journey mode
-							// (with SSC on) -- particles, May 25, 2 Reiwa
-							if (!Main.GameModeInfo.IsJourneyMode)
-							{
-								return true;
-							}
-							if (!args.Player.HasPermission(Permissions.journey_setspawnrate))
-							{
-								args.Player.SendErrorMessage("You don't have permission to modify the NPC spawn rate of the server!");
-								return true;
-							}
-							break;
-						}
-					default:
-						{
-							return true;
-						}
-				}
-			} else if (moduleId == (int)NetModulesTypes.CreativeUnlocksPlayerReport && Main.GameModeInfo.IsJourneyMode)
-			{
-				var unknownField = args.Data.ReadByte();
-
-				if (unknownField == 0) //this is required or something???
-				{
-					var itemId = args.Data.ReadUInt16();
-					var amount = args.Data.ReadUInt16();
-
-					var totalSacrificed = TShock.ResearchDatastore.SacrificeItem(itemId, amount, args.Player);
-
-					var response = NetCreativeUnlocksModule.SerializeItemSacrifice(itemId, totalSacrificed);
-					NetManager.Instance.Broadcast(response);
-				}
+				return true;
 			}
 
-			// As of 1.4.x.x, this is now used for more things:
-			//	NetCreativePowersModule
-			//	NetCreativePowerPermissionsModule
-			//	NetLiquidModule
-			//	NetParticlesModule
-			//	NetPingModule
-			//	NetTeleportPylonModule
-			//	NetTextModule
-			// I (particles) have disabled the original return here, which means that we need to
-			// handle this more. In the interm, this unbreaks parts of vanilla. Originally
-			// we just blocked this because it was a liquid exploit.
 			return false;
 		}
 
@@ -3627,7 +3512,7 @@ namespace TShockAPI
 			if (!args.Player.HasPermission(Permissions.startdd2))
 			{
 				TShock.Log.ConsoleDebug("GetDataHandlers / HandleOldOnesArmy rejected permissions {0}", args.Player.Name);
-				args.Player.SendErrorMessage("You don't have permission to start the Old One's Army event.");
+				args.Player.SendErrorMessage("You do not have permission to start the Old One's Army event.");
 				return true;
 			}
 
@@ -3958,7 +3843,7 @@ namespace TShockAPI
 			public bool Killed { get; internal set; }
 		}
 
-		public enum NetModulesTypes
+		public enum NetModuleType
 		{
 			Liquid,
 			Text,